+++ changelog.md
@@ -12,6 +12,20 @@
 	webman-amplifier.php
 
 
+## 1.5.5
+
+* **Fix**: Compatibility with WordPress 5.0
+
+### Files changed:
+
+	changelog.md
+	class-wm-amplifier.php
+	readme.txt
+	webman-amplifier.php
+	includes/metabox/class-metabox.php
+	includes/visual-editor/visual-editor.php
+
+
 ## 1.5.4
 
 * **Fix**: Compatibility with WooSidebars